made admin reqd, implementing authorizing

+++ blueprints/auth_bp.py
@@ -9,17 +9,17 @@ from datetime import timedelta
 
 bp_users = Blueprint("bp_users",__name__, url_prefix='/users')
 
-# Admin Searches all users (NEED JWT)
+# Search all users (Admin Only access)
 @bp_users.route('/')
 @jwt_required()
 def read_all_users():
-    # Need Admin Reqd
+    admin_only()
     stmt = db.select(User)
     users = db.session.scalars(stmt).all()
 
     return UserSchema(many = True, exclude = ['password']).dump(users)
 
-# Admin Searches for specific user ---(NEEDS AUTHORIZING-ADMIN ONLY ) + JWT
+# User can look at their own account details only
 @bp_users.route('/<int:id>')
 @jwt_required()
 def read_single_user(id):
@@ -31,12 +31,11 @@ def read_single_user(id):
     else:
         return {'Error': f'User ID {user_id} not found'}
 
-
+# User can edit at their own account including password
 @bp_users.route('/<int:id>', methods = ['PUT'])
 @jwt_required()
 def update_user(id):
     user_info = UserSchema(exclude=['id', 'admin_acc']).load(request.json)
-    
     user = db.session.query(User).filter_by(id=id).first()
 
     if user:
@@ -63,12 +62,13 @@ def update_user(id):
 def delete_user(id):
     stmt= db.select(User).filter_by(id = id)
     user= db.session.scalar(stmt)
-
-    if user:
-        owner_admin_authorize(user.id)
+    if user and user.admin_acc == False:
+        admin_only()
         db.session.delete(user)
         db.session.commit()
-        return {'Success': f'User ID {id} and all related content deleted'}
+        return {'Success': f'User ID {id} and all related Trips deleted'}
+    else:
+        return {'Error': 'You cannot delete an Admin'}
 
 
 # User Creates Acc
@@ -124,9 +124,18 @@ def login():
         return {"Error": "Email and Password need to be provided."},400
 
 
+def admin_only():
+    current_user_id = get_jwt_identity()
+    stmt = db.select(User).filter_by(id=current_user_id)
+    user = db.session.scalar(stmt)
+    if not (user and user.admin_acc):
+        abort(401, description="You must be an admin to access")
+
+
 def owner_admin_authorize(owner_id):
     current_user_id = get_jwt_identity()
     stmt = db.select(User).filter_by(id=current_user_id)
     user = db.session.scalar(stmt)
     if not (user and (user.admin_acc or current_user_id == owner_id)):
-        abort(401, description="You must be an admin or user")
+        abort(401, description="You must be an admin or user to access")
+
